13. Throwable Item 상호 작용 구현 & AI Battle 콤보 공격 수정.
- 플레이어 입력을 통해 Throwable Item과의 상호 작용하는 기능을 구현.(집고 던지기)
- AI Battle의 경우 Player Battle과 비슷하게 각기 하나의 콤보 공격을 사용할 수 있도록 수정.

+++ Scripts/HealthFoodItem.cs
@@ -8,7 +8,7 @@ public class HealthFoodItem : InteractableItem
 
     void Start()
     {
-        itemType = (int)ItemType.FOOD;
+        itemType = ItemType.FOOD;
     }
 
     protected override void Awake()
